+++ internal/converter/internal/otelcolconvert/converter_servicegraphconnector.go
@@ -64,5 +64,6 @@ func toServicegraphConnector(state *State, id component.InstanceID, cfg *service
 		Output: &otelcol.ConsumerArguments{
 			Metrics: ToTokenizedConsumers(nextMetrics),
 		},
+		VirtualNodePeerAttributes: cfg.VirtualNodePeerAttributes,
 	}
 }