+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Main (unreleased)
 
 - (_Public preview_) Add native histogram support to `otelcol.receiver.prometheus`. (@wildum)
 
+- Add the virtual_node_peer_attributes setting to `otelcol.connector.servicegraph`. (@hainenber)
+
 ### Bugfixes
 
 - Fix panic when component ID contains `/` in `otelcomponent.MustNewType(ID)`.(@qclaogui)